requirements.txt
Pillow
torchvision
numpy
